requirements.txt
litellm>=1.30.0
tavily-python>=0.3.0
composio>=1.0.0rc2
mcp>=1.0.0
pypdf>=4.0.0
reportlab>=4.1.0
apscheduler>=3.10.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
filelock>=3.13.0
pytest>=8.0.0
pytest-mock>=3.12.0
requests>=2.31.0
